Fixes #128 - Rework intuition buff requirement checks
* Reworked how intuition buff is computed.
When a fish requires intuition, and the pre-reqs are up at different times, the fish will only be considered available when all the pre-reqs have appeared overlapping or adjacent to each other.
This should fix issues with Warden of the Seven Hues.

* Changed "predators" key from object to list.
In order for the new intuition checks to function optimally, the list of pre-reqs must be ordered as expected. Javascript was automatically ordering the keys when they appeared to be numbers, which was not the desired behavior.