(fixes #150) detect inituition fish availability changes (#153)
* (fixes #150) detect inituition fish availability changes

When the layout updates subEntries (intuition fish) it does not capture
whether the subEntry availability has changed. This means intuition fish
availability changes don't provoke a layout sort. Without a layout sort
the "fish-availability-transitioning" class won't get removed, so the
text for the time remaining remains set of 0 opacity.

* Revert "(fixes #150) detect inituition fish availability changes"

This reverts commit 2919b88286c523ef7c3883402ffa906f27d1b7a8.

* Remove fish-availability-transitioning instead

This functionality was not working as intended so rather than forcing an
expensive sort on intuition fish availability changes instead we just
remove the `fish-availability-transitioning` class altogether to resolve
(#150)